+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### StreamChatUI
 ### 🐞 Fixed
 - Fix composer content not cleared when draft deleted event is triggered [#3626](https://github.com/GetStream/stream-chat-swift/pull/3626)
+- Set `ColorPalette.text` to `titleLabel` in `ChatChannelListItemView` [#3629](https://github.com/GetStream/stream-chat-swift/pull/3629)
 
 # [4.74.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.74.0)
 _March 14, 2025_